requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.119.1
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.0
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.12.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
sentry-sdk==2.58.0
sniffio==1.3.1
starlette==0.48.0
supabase==2.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
upstash-redis==1.7.0
uvicorn==0.38.0
vercel==0.3.2
vercel-sandbox==0.0.2
vercel-sdk==0.0.8
